coalesce the various gurka access tests into one
